+++ src/main/java/edu/illinois/library/cantaloupe/util/SocketUtils.java
@@ -2,7 +2,9 @@ package edu.illinois.library.cantaloupe.util;
 
 import java.io.IOException;
 import java.net.ServerSocket;
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 public final class SocketUtils {
@@ -36,6 +38,27 @@ public final class SocketUtils {
         }
 
         return ports;
+    }
+
+    public static int getUsedPort() {
+        return getUsedPorts(1)[0];
+    }
+
+    public static int[] getUsedPorts(int howMany) {
+        final List<Integer> ports = new ArrayList<>(65535);
+        int numFound = 0;
+        for (int port = 1; port < 65535; port++) {
+            if (numFound >= howMany) {
+                break;
+            }
+            //noinspection EmptyTryBlock
+            try (ServerSocket socket = new ServerSocket(port)) {
+            } catch (IOException e) {
+                ports.add(port);
+                numFound++;
+            }
+        }
+        return ports.stream().mapToInt(i -> i).toArray();
     }
 
     private SocketUtils() {}